Disable Turbolinks caching
Caching of Turbolinks visits causes more harm then benefits. For example we need to take care of tearing down all third party JS widgets (like select2, jQueryUI) on our own. Also lots of our JS is written in a way that it needs a fresh copy of the body to work with.