Allow refreshing rates for unshipped complete orders
When an order is complete, but not shipped yet, admins need to be able
to refresh rates for a shipment.

This is especially the case when splitting shipments to a new location.
This will create a new shipment with no rates, and refreshing them is the
only way to get any.

I added an expectation for a new shipment to be generated with a selected
shipping rate to test this behaviour.