Refactor redirecting on invalid order state
In the frontend checkout process, there is two methods that handle
validating the correct order state (or desired state):
`#set_state_if_present` and `#ensure_valid_state`. Both redirect
under certain conditions.

This commit changes that behaviour such that:
`#set_state_if_present` sets the current order's state to
the desired state so that partials are rendered correctly.
This method now also clears the parameter if the desired state is
not a possible future state of the order.

All redirects are moved to `#ensure_valid_state`. If any redirects happen, the
method returns so that no double render errors are possible.